Add files via upload
This is the output file from the image text extraction. Since the images were not super clear, and also due to the different background colors in the images, the extraction was of course not perfect. It was pretty good though.
I have corrected the first two images output but i'm tired of reading every line and trying to correct (lol it's almost as though i'm typing the whole damn thing, which is what we were trying to avoid in the first place) so i'll continue later and finish before the next meetup. 
Anyways i just wanted to show you guys how it went.